+++ app/src/main/java/com/example/appdevelopment/ui/screens/cameraView/CameraScreen.kt
@@ -1,6 +1,5 @@
 package com.example.appdevelopment.ui.screens.cameraView
 
-
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
@@ -28,18 +27,15 @@ import com.example.appdevelopment.data.AuthLogic
 import com.example.appdevelopment.mockData.posts.posts
 //import com.example.appdevelopment.ui.components.BottomHomeBar
 import com.example.appdevelopment.ui.components.TopHomeBar
+import com.example.appdevelopment.ui.layout.Scaffoldlayout
 import com.example.appdevelopment.ui.screens.feedView.FeedScreen
 import com.example.appdevelopment.ui.screens.feedView.PostList
-import com.example.appdevelopment.ui.screens.feedView.Scaffoldlayout
-
 
 @ExperimentalMaterial3Api
 @Composable
-
-
 fun CameraScreen(navController: NavController, authLogic: AuthLogic?) {
 
-    Scaffoldcamlayout(navController = navController, "Sailboat")
+    Scaffoldlayout(navController = navController, "Sailboat") { CameraX() }
     
    Button(
        onClick = {
@@ -55,83 +51,7 @@ fun CameraScreen(navController: NavController, authLogic: AuthLogic?) {
    }
 }
 
-@ExperimentalMaterial3Api
 @Composable
-fun Scaffoldcamlayout(navController: NavController, dailyWord: String) {
-    androidx.compose.material3.Scaffold(
-        topBar = {
-            CenterAlignedTopAppBar(
-                title = {
-                    androidx.compose.material3.Text(
-                        text = dailyWord,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                },
-                navigationIcon = {
-                    IconButton(onClick = { navController.navigate(Screen.Profile.route) }) {
-                        Icon(
-                            Icons.Default.Person,
-                            contentDescription = "Localized description",
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                }
-            )
-        },
-        bottomBar = {
-            BottomAppBar(
-                Modifier
-                    .fillMaxWidth()
-                    .wrapContentHeight(),
-                backgroundColor = Color.Black,
-            ) {
-                Row(
-                    horizontalArrangement = Arrangement.SpaceEvenly,
-                    modifier = Modifier.fillMaxWidth(1f)
-                ) {
-                    IconButton(onClick = { navController.navigate(Screen.Feed.route) }) {
-                        Icon(
-                            painter = painterResource(
-                                id = R.drawable.ic_baseline_menu_24
-                            ),
-                            contentDescription = null,
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                    IconButton(onClick = { navController.navigate(Screen.Camera.route) }) {
-                        Icon(
-                            painter = painterResource(
-                                id = R.drawable.ic_baseline_photo_camera_24
-                            ),
-                            contentDescription = null,
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                    IconButton(onClick = { navController.navigate(Screen.Leaderboards.route) }) {
-                        Icon(
-                            painter = painterResource(
-                                id = R.drawable.ic_baseline_leaderboard_24
-                            ),
-                            contentDescription = null,
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                }
-            }
-        }
-    ) {
-        Text(text = "Hej CameraX")
-    }
+fun CameraX() {
+    Text(text = "Hej CameraX")
 }
-
-
-@ExperimentalMaterial3Api
-@Preview
-@Composable
-fun CameraScreenPreview() {
-    //CameraScreen(navController = NavController())
-}
-
-
